Bring Lobby test coverage to 50%+ Fixes #1203
